fixed: Seg fault in xmbctex on 64bit linux.
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/linuxport@21469 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ XBMC/tools/XBMCTex/sha.c
@@ -5,7 +5,7 @@
 #include "stdint_win.h"
 #endif
 
-typedef unsigned int u32;
+typedef unsigned long u32;
 typedef unsigned char u8;
 
 #ifdef _LINUX
